Use built-in actions for SyncedFolders and WaitForCommunicator
